initial codes

+++ src/main/java/shblock/interactivecorporea/client/requestinghalo/animation/AbstractAnimation.java
@@ -1,13 +1,13 @@
 package shblock.interactivecorporea.client.requestinghalo.animation;
 
 public class AbstractAnimation {
-  protected float time = 1F;
-  private final float speed;
+  protected double time = 1F;
+  private final double speed;
 
   /**
    * @param length The ticks it takes for this animation to complete
    */
-  public AbstractAnimation(float length) {
+  public AbstractAnimation(double length) {
     speed = 1F / length;
   }
 
@@ -15,7 +15,7 @@ public class AbstractAnimation {
    * Update the animation (Called every frame)
    * @return If the animation should be removed (If the animation is done)
    */
-  public boolean update(float dt) {
+  public boolean update(double dt) {
     time -= speed * dt;
     if (time < 0F) {
       time = 0F;
