+++ src/main/java/shblock/interactivecorporea/client/requestinghalo/animation/AnimationFadeOut.java
@@ -1,7 +1,7 @@
 package shblock.interactivecorporea.client.requestinghalo.animation;
 
 public class AnimationFadeOut extends AbstractAnimation {
-  public AnimationFadeOut(float length) {
+  public AnimationFadeOut(double length) {
     super(length);
   }
 
@@ -10,7 +10,7 @@ public class AnimationFadeOut extends AbstractAnimation {
   }
 
   public int getAlpha() {
-    float p = (float) Math.sin(Math.PI * time * .5F);
+    double p = Math.sin(Math.PI * time * .5F);
     return (int) (p * 255);
   }
 }
